+++ src/savedsearches.cpp
@@ -109,7 +109,7 @@ void SavedSearches::retrieveFromStorage( QSettings& settings )
     if ( settings.contains( "SavedSearches/version" ) ) {
         // Unserialise the "new style" stored history
         settings.beginGroup( "SavedSearches" );
-        if ( settings.value( "version" ) == SAVEDSEARCHES_VERSION ) {
+        if ( settings.value( "version" ).toInt() == SAVEDSEARCHES_VERSION ) {
             int size = settings.beginReadArray( "searchHistory" );
             for (int i = 0; i < size; ++i) {
                 settings.setArrayIndex(i);
